Move BuildingUpgrade class to its own file
