+++ Assets/Scripts/ScriptableObjects/Building.cs
@@ -1,6 +1,5 @@
-using System;
+using Classes;
 using UnityEngine;
-using Resources = Classes.Resources;
 
 namespace ScriptableObjects
 {
@@ -10,14 +9,5 @@ namespace ScriptableObjects
         public string buildingName;
         public string description;
         public BuildingUpgrade[] upgrades;
-    }
-
-    [Serializable]
-    public class BuildingUpgrade
-    {
-        public int level; // this should be always incremented by 1
-        public Resources cost;
-        public Resources output;
-        public GameObject model;
     }
 }
